+++ src/oersted/biotsavart.py
@@ -5,7 +5,7 @@ from numpy.typing import NDArray
 
 # Create bindings for calculation engine written in Rust
 from ._oersted import (
-    b_current_point_direct,
+    h_current_point_direct,
     h_current_point_octree,
     h_current_tet4_direct,
     h_current_tet4_octree,
@@ -28,8 +28,18 @@ def b_field(
     solver: DirectSolver | OctreeSolver | None = None,
 ) -> NDArray[float64]:
     """Compute the magnetic flux density at a collection of target points using the
-    specific source mesh and solver options, assuming the target points are
-    in free space
+    specific source mesh and solver options, assuming the target points are in free
+    space
+
+    Args:
+        source: mesh to use as the field source
+        j_density: (A/m^2) (N,3) array of current density vectors at each of the
+            element centroids
+        targets: (m) (N,3) array of target point positions in 3D space
+        solver: selects the solution settings
+
+    Returns:
+        (T) (N,3) array of magnetic flux density (B) vectors at each target position
     """
     return MU0 * h_field(source, j_density, targets, solver)
 
@@ -39,9 +49,21 @@ def h_field(
     j_density: NDArray[float64],
     targets: NDArray[float64],
     solver: DirectSolver | OctreeSolver | None = None,
+    edge: bool = False,
 ) -> NDArray[float64]:
     """Compute the magnetic field strength at a collection of target points using
-    a current-carrying source mesh."""
+    a current-carrying source mesh.
+
+    Args:
+        source: mesh to use as the field source
+        j_density: (A/m^2) (N,3) array of current density vectors at each of the
+            element centroids
+        targets: (m) (N,3) array of target point positions in 3D space
+        solver: selects the solution settings
+
+    Returns:
+        (T) (N,3) array of magnetic field strength (H) vectors at each target position
+    """
 
     if solver is None:
         solver = DirectSolver()
@@ -54,7 +76,7 @@ def h_field(
         src_vol = ascontiguousarray(source.volumes, dtype=float64)
 
         if isinstance(solver, DirectSolver):
-            return (1.0 / MU0) * b_current_point_direct(
+            return h_current_point_direct(
                 src_pts, src_vol, j_density, tgt_pts, solver.n_threads
             )
 
@@ -79,7 +101,7 @@ def h_field(
         src_connectivity = ascontiguousarray(source.connectivity, dtype=uint32)
         if isinstance(solver, DirectSolver):
             return h_current_tet4_direct(
-                src_nodes, src_connectivity, j_density, tgt_pts, solver.n_threads
+                src_nodes, src_connectivity, j_density, tgt_pts, solver.n_threads, edge
             )
 
         elif isinstance(solver, OctreeSolver):
@@ -110,7 +132,18 @@ def h_mag(
     targets: NDArray[float64],
     solver: DirectSolver | OctreeSolver | None = None,
 ) -> NDArray[float64]:
-    """Compute the magnetic field strength using a magnetized mesh as the source"""
+    """Compute the magnetic field strength using a magnetized mesh as the source
+
+    Args:
+        source: mesh to use as the field source
+        m_field: (A/m) (N,3) array of magnetization field vectors at each of the
+            element centroids
+        targets: (m) (N,3) array of target point positions in 3D space
+        solver: selects the solution settings
+
+    Returns:
+        (T) (N,3) array of magnetic field strength (H) vectors at each target position
+    """
 
     if solver is None:
         solver = DirectSolver()
@@ -171,6 +204,7 @@ def h_mag(
                 leaf_threshold,
                 solver.n_threads,
                 use_octree,
+                solver.edge,
             )
 
         elif isinstance(solver, OctreeSolver):
@@ -184,6 +218,7 @@ def h_mag(
                 solver.leaf_threshold,
                 solver.n_threads,
                 use_octree,
+                solver.edge,
             )
 
         else: